chore(ui-kit): renamed review coaching session modal for consistency, removed unused imports
