+++ packages/ui-kit/tests/review-coachingsession-modal.test.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
 import { render, screen, fireEvent, act } from '@testing-library/react';
 import { describe, it, expect, vi } from 'vitest';
-import { ReviewCoachingSessionModal } from '../lib/components/review-coachingsession-modal';
+import { ReviewCoachingSessionModal } from '../lib/components/review-coaching-session-modal';
 
 // Mock dependencies
 vi.mock('@maany_shr/e-class-translations', () => ({
